requirements.txt
Pillow>=9.4.0,<9.5.0
